feat: add lua functions and WSL integration
This commit adds the ability to use lua functions for
creating code images, which makes lua shortcut mappings
a little bit easier.

Support for simultaneous clipboard and file code
images has been added.

WSL2 is detected and if an image shall be put onto the
clipboard, a Powershell helper script is executed that
puts the image onto the Windows clipboard.

The lua module has been renamed to 'nvim-silicon' to
match the github repo name. An update recommendation
is displayed to the user after a code image has been
generated.